+++ src/hooks/mutations/useUpdateCredit.tsx
@@ -42,7 +42,7 @@ const useUpdateCredit = () => {
 			refetchQueries: [
 				{
 					query: QUERY_PROFILE,
-					variables: { id: userId, author: userId, last: 5 },
+					variables: { id: userId, author: userId, lastCredits: 5 },
 				},
 			],
 		});